11월 11일

+++ src/main/java/org/zerock/controller/mamber/MemberController.java
@@ -45,26 +45,44 @@ public class MemberController {
 	}
 	
 	@PostMapping("modify")
-	public String modify(MemberDto member, RedirectAttributes rttr) {
-		int cnt = service.modify(member);
+	public String modify(MemberDto member, String oldPassword, RedirectAttributes rttr) {
+		
+		MemberDto oldmember = service.getById(member.getId());
 		
 		rttr.addAttribute("id", member.getId());
-		if (cnt == 1) {
-			rttr.addFlashAttribute("message", "회원 정보가 수정되었습니다.");
-			return "redirect:/member/info";
+		if(oldmember.getPassword().equals(oldPassword)) {
+			// 기존 암호가 맞으면 회원 정보 수정 
+			int cnt = service.modify(member);
+			
+			if (cnt == 1) {
+				rttr.addFlashAttribute("message", "회원 정보가 수정되었습니다.");
+				return "redirect:/member/info";
+			} else {
+				rttr.addFlashAttribute("message", "회원 정보가 수정되지 않았습니다.");
+				return "redirect:/member/modify";
+			}
 		} else {
-			rttr.addFlashAttribute("message", "회원 정보가 수정되지 않았습니다.");
+			rttr.addFlashAttribute("message", "암호가 일치하지 않습니다.");
 			return "redirect:/member/modify";
 		}
+		
 	}
 	
 	@PostMapping("remove")
-	public String remove(String id, RedirectAttributes rttr) {
-		int cnt = service.remove(id);
-		
-		rttr.addFlashAttribute("message", "회원 정보가 삭제되었습니다.");
-		
-		return "redirect:/member/list";
+	public String remove(String id, String oldPassword, RedirectAttributes rttr) {
+		MemberDto oldmember = service.getById(id);
+		if (oldmember.getPassword().equals(oldPassword)) {
+			int cnt = service.remove(id);
+			
+			rttr.addFlashAttribute("message", "회원 탈퇴하였습니다.");
+			
+			return "redirect:/member/list";
+			
+		} else {
+			rttr.addAttribute("id", id);
+			rttr.addFlashAttribute("message", "암호가 일치하지 않습니다.");
+			return "redirect:/member/modify";
+		}
 
 	}
 }
